ProxyShuffleManager cannot be dedupped, so undoing
- WIP directory renames
