+++ shims/spark302/pom.xml
@@ -109,8 +109,8 @@
                                 <phase>generate-sources</phase>
                                 <configuration>
                                     <sources>
-                                        <source>${spark-rapids.sql-plugin.root}/src/main/spark301+all/scala</source>
-                                        <source>${spark-rapids.sql-plugin.root}/src/main/spark301+311-all/scala</source>
+                                        <source>${spark-rapids.sql-plugin.root}/src/main/301until320-all/scala</source>
+                                        <source>${spark-rapids.sql-plugin.root}/src/main/301until311-all/scala</source>
                                     </sources>
                                 </configuration>
                             </execution>
