+++ shims/spark311cdh/src/main/scala/com/nvidia/spark/rapids/spark311cdh/RapidsShuffleManager.scala
@@ -17,10 +17,11 @@
 package com.nvidia.spark.rapids.spark311cdh
 
 import org.apache.spark.SparkConf
-import org.apache.spark.rapids.shims.v2.ProxyRapidsShuffleInternalManager
+import org.apache.spark.sql.rapids.shims.spark311cdh.ProxyRapidsShuffleInternalManager
 
 /** A shuffle manager optimized for the RAPIDS Plugin for Apache Spark. */
 sealed class RapidsShuffleManager(
     conf: SparkConf,
     isDriver: Boolean) extends ProxyRapidsShuffleInternalManager(conf, isDriver) {
 }
+
