+++ shims/spark320/src/main/scala/com/nvidia/spark/rapids/spark320/RapidsShuffleManager.scala
@@ -17,7 +17,7 @@
 package com.nvidia.spark.rapids.spark320
 
 import org.apache.spark.SparkConf
-import org.apache.spark.rapids.shims.v2.ProxyRapidsShuffleInternalManager
+import org.apache.spark.sql.rapids.shims.spark320.ProxyRapidsShuffleInternalManager
 
 /** A shuffle manager optimized for the RAPIDS Plugin for Apache Spark. */
 sealed class RapidsShuffleManager(
