+++ sql-plugin/src/main/scala/org/apache/spark/sql/rapids/RapidsShuffleInternalManagerBase.scala
@@ -427,3 +427,41 @@ abstract class RapidsShuffleInternalManagerBase(conf: SparkConf, val isDriver: B
 trait VisibleShuffleManager extends ShuffleManager {
   def isDriver: Boolean
 }
+
+abstract class ProxyRapidsShuffleInternalManagerBase(
+    conf: SparkConf,
+    override val isDriver: Boolean
+) extends VisibleShuffleManager with Proxy {
+
+  // touched in the plugin code after the shim initialization
+  // is complete
+  override lazy val self: VisibleShuffleManager =
+    ShimLoader.newInternalShuffleManager(conf, isDriver)
+
+
+  //
+  // Signatures unchanged since 3.0.1 follow
+  //
+
+  override def getWriter[K, V](
+      handle: ShuffleHandle,
+      mapId: Long,
+      context: TaskContext,
+      metrics: ShuffleWriteMetricsReporter
+  ): ShuffleWriter[K, V] = {
+    self.getWriter(handle, mapId, context, metrics)
+  }
+
+  override def registerShuffle[K, V, C](
+      shuffleId: Int,
+      dependency: ShuffleDependency[K, V, C]
+  ): ShuffleHandle = {
+    self.registerShuffle(shuffleId, dependency)
+  }
+
+  override def unregisterShuffle(shuffleId: Int): Boolean = self.unregisterShuffle(shuffleId)
+
+  override def shuffleBlockResolver: ShuffleBlockResolver = self.shuffleBlockResolver
+
+  override def stop(): Unit = self.stop()
+}
